### 2026-02-26 #10120 @danielskatz
👋 @vissarion - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. In addition, your paper is missing many of the currently required sections, as mentioned in a comment above. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. An editorialbot command needs to be the first entry in a new comment.
