### 2026-03-05 06e04c6
Rewrite in new JOSS format
