+++ joss_paper/paper.md
@@ -16,19 +16,14 @@ authors:
     orcid: 0000-0002-0780-666X
     affiliation: 2
   - name: Barend Gehrels
-    orcid: 0000-0000-0000-0000
     affiliation: 1
   - name: Menelaos Karavelas
-    orcid: 0000-0000-0000-0000
     affiliation: 1
   - name: Bruno Lalande
-    orcid: 0000-0000-0000-0000
     affiliation: 1
   - name: Mateusz Łoskot
-    orcid: 0000-0000-0000-0000
     affiliation: 1
   - name: Adam Wulkiewicz
-    orcid: 0000-0000-0000-0000
     affiliation: 1
 affiliations:
   - name: Boost C++ Libraries
@@ -45,13 +40,13 @@ bibliography: paper.bib
 
 Computational geometry underpins numerous scientific and engineering domains, including GIS, robotics, computer graphics, CAD, and astronomy. Foundational algorithms such as convex hulls, intersection tests, and point-in-polygon queries are covered in standard texts [@deBerg:2008]; however, robust and efficient implementations are crucial for real-world applications where numerical robustness and performance matter.
 
-**Boost.Geometry**[^*] is an open-source C++ library providing generic, extensible, and standards-compliant geometric computation. It defines geometry *concepts* and implements algorithms in a dimension- and coordinate-system-agnostic manner. Since its inclusion in Boost (2011), it has become a widely used component in C++ scientific and industrial software. For example, MySQL uses Boost.Geometry as the geometry engine for spatial operations [@MySQLGIS:2014].
+`Boost.Geometry`[^*] is an open-source C++ library providing generic, extensible, and standards-compliant geometric computation. It defines geometry *concepts* and implements algorithms in a dimension- and coordinate-system-agnostic manner. Since its inclusion in Boost (2011), it has become a widely used component in C++ scientific and industrial software.
 
 [^*]: Authors are listed in alphabetical order by surname.
 
-# Statement of Need
+# Statement of need
 
-Boost.Geometry fulfills the need for a general-purpose, high-performance geometry engine in C++ with:
+`Boost.Geometry` fulfills the need for a general-purpose, high-performance geometry engine in C++ with:
 
 - A generic programming architecture via concepts, type traits, and template metaprogramming
 - Strategy-based dispatch enabling coordinate-system-specific algorithms
@@ -59,6 +54,8 @@ Boost.Geometry fulfills the need for a general-purpose, high-performance geometr
 - OGC-compliant geometry types and spatial predicates [@OGC:2011]
 - Header-only distribution, simplifying deployment and integration
 - Support for adapting user-defined geometry types
+
+# State of the field 
 
 A similar library is GEOS [@GEOS:2025], a C++ port of the Java Topology Suite (JTS) that implements the OGC Simple Features standard for GIS applications and serves as the core geometry engine for major geospatial tools like PostGIS [@PostGIS:2025].
 
@@ -68,27 +65,52 @@ Coordinate transformations and CRS handling are commonly performed using PROJ [@
 
 GeographicLib [@GeographicLib:2013] and Karney's work on geodesics [@Karney:2013] provide state-of-the-art algorithms for ellipsoidal geodesic problems.
 
-# Design and Features
+# Software design
 
-Boost.Geometry provides:
+`Boost.Geometry` provides:
 
 - **Geometry models:** points, boxes, rings, polygons, multi-geometries
 - **Algorithms:** area, distance, length, centroid, convex hull, within, intersects, union, intersection, simplify, buffer, transformations and projections
 - **Spatial indexing:** R-tree implementation supporting spatial and nearest-neighbor queries
 
[…]
