+++ joss_paper/paper.bib
@@ -41,8 +41,7 @@
   author = {{GEOS contributors}},
   organization = {Open Source Geospatial Foundation},
   year = {2025},
-  url = {https://libgeos.org/},
-  doi = {10.5281/zenodo.11396894}
+  url = {https://libgeos.org/}
 }
 
 @article{Karney:2013,
@@ -62,7 +61,7 @@
   author={Zhao, Jun},
   year={2014},
   note={MySQL Developer Blog},
-  url={https://blogs.oracle.com/mysql/post/mysql-57-gis-enhancements},
+  url={https://dev.mysql.com/blog-archive/making-use-of-boost-geometry-in-mysql-gis/},
   urldate={2026-01-27}
 }
 
@@ -100,7 +99,7 @@
 }
 
 @misc{libspatialindex:2024,
-  title={libspatialindex - R-tree spatial index library},
+  title={libspatialindex - {R}-tree spatial index library},
   author={Libspatialindex Contributors},
   year={2024},
   url={https://libspatialindex.org/},
@@ -137,6 +136,87 @@ DOI = {10.5194/isprs-archives-XLVI-4-W2-2021-1-2021}
   year={2004},
   organization={Springer},
   doi={10.1016/J.COMGEO.2007.06.003}
+}
+
+@inproceedings{Georgiadis:2025,
+  author    = {Georgiadis, Thanasis and Michalopoulos, Achilleas and Dimitropoulos, Dimitris and Tsitsigkos, Dimitris and Mamoulis, Nikos},
+  title     = {Hecatoncheir: Scaling up and out spatial data management},
+  booktitle = {Proceedings of the 33rd ACM International Conference on Advances in Geographic Information Systems (SIGSPATIAL '25)},
+  pages     = {844--847},
+  year      = {2025},
+  publisher = {ACM},
+  doi       = {10.1145/3748636.3762796}
+}
+
+@article{Bucelli:2024,
+  author = {Bucelli, Michele},
+  title = {The lifex Library Version 2.0},
+  year = {2025},
+  issue_date = {December 2025},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  volume = {51},
+  number = {4},
+  issn = {0098-3500},
+  url = {https://doi.org/10.1145/3748817},
+  doi = {10.1145/3748817},
+  journal = {ACM Trans. Math. Softw.},
+  month = dec,
+  articleno = {28},
+  numpages = {10},
+  keywords = {High-Performance Computing, Finite elements, Numerical simulations, Multiphysics problems}
+}
+
+@inproceedings{Richoux:2022,
+  author    = {Richoux, Florian},
+  title     = {Terrain analysis in StarCraft 1 and 2 as combinatorial optimization},
+  booktitle = {2022 IEEE Congress on Evolutionary Computation (CEC)},
+  year      = {2022},
+  publisher = {IEEE},
+  doi       = {10.1109/CEC55065.2022.9870230}
+}
+
+@inproceedings{Ashtekar:2023,
+  author    = {Ashtekar, Vyankatesh and Dutta, Ashish},
+  title     = {Push Recovery Control of a Bipedal Robot Standing on Two Offset Planes in Double Leg Stance},
+  booktitle = {Proceedings of the 2023 6th International Conference on Advances in Robotics (AIR '23)},
+  articleno = {50},
+  pages     = {1--8},
+  year      = {2023},
+  publisher = {ACM},
+  doi       = {10.1145/3610419.3610469}
+}
+
+@article{Vermeulen:2017,
+  author    = {Vermeulen, Jordi L. and Hillebrand, Arne and Geraerts, Roland},
[…]
