### 2026-03-16 #10120 @hugoledoux
Very happy to do this one: I have used Boost.Geometry for a project in the past also
