### 2026-03-27 dd01797
Prevent FP contractions to FMA in index/detail/algorithms/content.hpp to fix #1452 on GCC with x86-64-v3.
