### 2026-03-30 #10328 @editorialbot
[REVIEW]: Boost.Geometry: A Generic C++ Geometry Library
**Submitting author:** @vissarion (<a href="http://orcid.org/0000-0002-0780-666X">Vissarion Fisikopoulos</a>)
**Repository:** https://github.com/boostorg/geometry
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** v1.90.0
