### 2026-04-07 7e9a531
Skip unnecessary, error-introducing longitude normalization on antimeridian-crossing box.
