### 2026-04-19 951e9fe
Drop varray implementation in favor of Boost.Container static_vector.
