### 2026-04-19 #10328 @ioannisman
@hugoledoux I will write here any feedback to @vissarion. I hope my understanding of how the review process and communication should take place is correct.

**Feedback on the paper**:
The paper contains all required sections and is overall well written. I found the Software Design and Research Impact Statement sections to be the strongest parts. Considering the review checklist, my main suggestion would be to improve how the paper positions the software, especially in the following two sections.

- The **Statement of need** already explains the technical need for the software and clearly lists the main capabilities of the library. However, it could say more explicitly who the software is for in practice and what concrete user needs it is meant to address.
- The **State of the field** already gives a useful overview of related software and what these tools provide. However, it could compare them more directly to Boost.Geometry and explain more clearly what added value this library brings compared to these existing alternatives.

Minor suggestion: in the **Research impact statement**, the library is likely applicable in many more contexts than those currently listed (?). If so, it may be helpful to make clear that these are indicative examples rather than a complete overview.
