### 2026-04-20 fe0f2f8
fix(rtree): clamp content differences to prevent FMA-induced numerical instability
Might fix #1452
