+++ include/boost/geometry/index/detail/rtree/rstar/choose_next_node.hpp
@@ -117,7 +117,7 @@ private:
 
             // areas difference
             content_type content = index::detail::content(box_exp);
-            content_type content_diff = content - index::detail::content(ch_i.first);
+            content_type content_diff = index::detail::content_diff(box_exp, ch_i.first);
 
             children_contents[i].set(i, content, content_diff);
 
@@ -246,7 +246,7 @@ private:
 
             // areas difference
             content_type content = index::detail::content(box_exp);
-            content_type content_diff = content - index::detail::content(ch_i.first);
+            content_type content_diff = index::detail::content_diff(box_exp, ch_i.first);
 
             // update the result
             if ( content_diff < smallest_content_diff ||
