### 2026-04-21 #10328 @hugoledoux
@ioannisman yes this is the correct way to address the paper. For the checklist, check the ones you see are fine, and if there are issues that need to be addressed you can also create a PR or an issue in the boost.geom repository and start a discussion there (if you can then put the link here in this issue it's helpful).

@vissarion I agree with the comments made above for the paper, if you could improve this slightly the paper would be stronger IMO
