### 2026-04-23 #10328 @kenohori
Hello! It's nice to see a submission for Boost.Geometry. I'm already somewhat familiar with the library from testing I did some years ago.

Regarding the paper, I basically agree with the points raised by @ioannisman:
* The **Statement of need** currently contains a description of the main features of Boost.Geometry, but it doesn't state the problems that the software is designed to solve, the target audience or its relation to other work.
* The **State of the field** refers to other related libraries, but these are not compared to Boost.Geometry.

Regarding the software and documentation, I've tested it with several of the examples in the documentation and the `example` folder in the repository. They all seem to work without problems. I've also run the automated tests through `b2` and I think it all works fine.

My only concern is that the `example` folder doesn't seem to be referred to in the documentation? I found it rather confusing to only include examples for Spatial Indexes in the documentation and the rest only provided as source code (without much of an explanation). Do the examples in the `example` folder cover all of the library?
