### 2026-04-23 83b2bc5
Use std::addressof over boost::addressof
