### 2026-04-23 7c4a076
Use std::addressof over boost::addressof
