### 2026-04-23 10289e5
Use C++11 explicitly deleted copy constructors over boost::noncopyable.
