### 2026-04-23 3f0bc5d
Use C++11 explicitly deleted copy constructors over boost::noncopyable.
