### 2026-04-24 08c7cdb
Remove dependencies on Boost.Array excluding examples that document adapted/boost_array.hpp
