+++ include/boost/geometry/util/series_expansion.hpp
@@ -30,11 +30,20 @@
 #ifndef BOOST_GEOMETRY_UTIL_SERIES_EXPANSION_HPP
 #define BOOST_GEOMETRY_UTIL_SERIES_EXPANSION_HPP
 
-#include <boost/array.hpp>
+#include <array>
+
 #include <boost/geometry/core/assert.hpp>
 #include <boost/geometry/util/math.hpp>
 
 namespace boost { namespace geometry { namespace series_expansion {
+
+    template <typename T, size_t N>
+    auto array_base_size_helper(std::array<T, N> const*) 
+        -> std::integral_constant<size_t, N>;
+
+    template <typename Coeffs>
+    constexpr size_t static_size = 
+        decltype(array_base_size_helper(static_cast<Coeffs*>(nullptr)))::value;
 
     /*
      Generate and evaluate the series expansion of the following integral
@@ -156,7 +165,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     template <typename Coeffs, typename CT>
     inline void evaluate_coeffs_A3(Coeffs &c, CT const& n)
     {
-        switch (int(Coeffs::static_size))
+        switch (int(static_size<Coeffs>))
         {
         case 0:
             break;
@@ -228,7 +237,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     {
         CT const eps2 = math::sqr(eps);
         CT d = eps;
-        switch (int(Coeffs::static_size) - 1)
+        switch (int(static_size<Coeffs>) - 1)
         {
         case 0:
             break;
@@ -328,7 +337,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     {
         CT const eps2 = math::sqr(eps);
         CT d = eps;
-        switch (int(Coeffs::static_size) - 1)
+        switch (int(static_size<Coeffs>) - 1)
         {
         case 0:
             break;
@@ -428,7 +437,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     {
         CT const eps2 = math::sqr(eps);
         CT d = eps;
-        switch (int(Coeffs::static_size) - 1)
+        switch (int(static_size<Coeffs>) - 1)
         {
         case 0:
             break;
@@ -525,7 +534,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     */
     template <size_t SeriesOrder, typename Coeffs, typename CT>
     inline void evaluate_coeffs_C3x(Coeffs &c, CT const& n) {
-        BOOST_GEOMETRY_ASSERT((Coeffs::static_size == (SeriesOrder * (SeriesOrder - 1)) / 2));
+        BOOST_GEOMETRY_ASSERT((static_size<Coeffs> == (SeriesOrder * (SeriesOrder - 1)) / 2));
 
         CT const n2 = math::sqr(n);
         switch (SeriesOrder)
@@ -648,10 +657,10 @@ namespace boost { namespace geometry { namespace series_expansion {
         size_t offset = 0;
 
         // i is the index of C3[i].
-        for (size_t i = 1; i < Coeffs1::static_size; ++i)
+        for (size_t i = 1; i < static_size<Coeffs1>; ++i)
         {
             // Order of polynomial in eps.
-            size_t m = Coeffs1::static_size - i;
+            size_t m = static_size<Coeffs1> - i;
             mult *= eps;
 
             coeffs1[i] = mult * math::horner_evaluate(eps, coeffs2.begin() + offset,
@@ -672,7 +681,7 @@ namespace boost { namespace geometry { namespace series_expansion {
     template <typename CT, typename Coeffs>
[…]
