### 2026-04-24 b5a0e53
[extensions] Remove dependency on Boost.StaticAssert.
