### 2026-04-24 18a521a
[extensions] Remove dependency on Boost.StaticAssert.
