+++ include/boost/geometry/extensions/gis/io/wkb/utility.hpp
@@ -20,7 +20,6 @@
 #include <type_traits>
 
 #include <boost/cstdint.hpp>
-#include <boost/static_assert.hpp>
 
 #include <boost/geometry/core/assert.hpp>
 
@@ -33,10 +32,10 @@ namespace boost { namespace geometry
 template <typename OutputIterator>
 bool hex2wkb(std::string const& hex, OutputIterator bytes)
 {
-    // Bytes can be only written to output iterator.
-    BOOST_STATIC_ASSERT((std::is_convertible<
+    static_assert(std::is_convertible<
         typename std::iterator_traits<OutputIterator>::iterator_category,
-        const std::output_iterator_tag&>::value));
+        const std::output_iterator_tag&>::value,
+        "Bytes can only be written to output iterator.");
 
     std::string::size_type const byte_size = 2;
     if (0 != hex.size() % byte_size)
@@ -64,10 +63,10 @@ bool hex2wkb(std::string const& hex, OutputIterator bytes)
 template <typename Iterator>
 bool wkb2hex(Iterator begin, Iterator end, std::string& hex)
 {
-    // Stream of bytes can only be passed using random access iterator.
-    BOOST_STATIC_ASSERT((std::is_convertible<
+    static_assert(std::is_convertible<
         typename std::iterator_traits<Iterator>::iterator_category,
-        const std::random_access_iterator_tag&>::value));
+        const std::random_access_iterator_tag&>::value,
+        "Stream of bytes can only be passed using random access iterator.");
 
     const char hexalpha[] = "0123456789ABCDEF";
     char hexbyte[3] = { 0 };
