### 2026-04-24 687281d
Remove dependency on Boost.StaticAssert.
