+++ include/boost/geometry/algorithms/detail/relate/areal_areal.hpp
@@ -206,9 +206,10 @@ private:
 template <typename Geometry1, typename Geometry2>
 struct areal_areal
 {
-    // check Linear / Areal
-    BOOST_STATIC_ASSERT(topological_dimension<Geometry1>::value == 2
-                     && topological_dimension<Geometry2>::value == 2);
+    // check Areal / Areal
+    static_assert(topological_dimension<Geometry1>::value == 2
+        && topological_dimension<Geometry2>::value == 2,
+        "Specialisation for areal / areal combination.");
 
     static const bool interruption_enabled = true;
 
