### 2026-04-24 35a7f48
Remove dependency on Boost.StaticAssert.
