+++ include/boost/geometry/algorithms/detail/relate/turns.hpp
@@ -247,7 +247,7 @@ struct less_other_multi_index
 template <std::size_t OpId, typename LessOp, typename Strategy>
 struct less
 {
-    BOOST_STATIC_ASSERT(OpId < 2);
+    static_assert(OpId < 2, "Only 0 and 1 are valid OpIds.");
 
     template <typename Turn>
     static inline bool use_fraction(Turn const& left, Turn const& right)
