+++ include/boost/geometry/srs/projections/impl/pj_gridinfo.hpp
@@ -43,16 +43,16 @@
 #define BOOST_GEOMETRY_SRS_PROJECTIONS_IMPL_PJ_GRIDINFO_HPP
 
 
+#include <algorithm>
+#icnlude <cstdint>
+#include <string>
+#include <vector>
+
 #include <boost/geometry/core/assert.hpp>
 #include <boost/geometry/util/math.hpp>
 
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/algorithm/string/trim.hpp>
-#include <boost/cstdint.hpp>
-
-#include <algorithm>
-#include <string>
-#include <vector>
 
 
 namespace boost { namespace geometry { namespace projections
@@ -106,7 +106,7 @@ struct pj_ctable
 {
     struct lp_t  { double lam, phi; };
     struct flp_t { float lam, phi; };
-    struct ilp_t { boost::int32_t lam, phi; };
+    struct ilp_t { std::int32_t lam, phi; };
 
     std::string id;         // ascii info
     lp_t ll;                // lower left corner coordinates
@@ -127,7 +127,7 @@ struct pj_ctable
 struct pj_gi_load
 {
     enum format_t { missing = 0, ntv1, ntv2, gtx, ctable, ctable2 };
-    typedef boost::long_long_type offset_t;
+    typedef std::intmax_t offset_t;
 
     explicit pj_gi_load(std::string const& gname = "",
                         format_t f = missing,
@@ -277,7 +277,7 @@ inline bool pj_gridinfo_load_ntv1(IStream & is, pj_gi_load & gi)
     std::vector<double> row_buf(r_size);
     gi.ct.cvs.resize(gi.ct.lim.lam * gi.ct.lim.phi);
 
-    for (boost::int32_t row = 0; row < gi.ct.lim.phi; row++ )
+    for (std::int32_t row = 0; row < gi.ct.lim.phi; row++ )
     {
         is.read(reinterpret_cast<char*>(&row_buf[0]), ch_size);
 
@@ -291,7 +291,7 @@ inline bool pj_gridinfo_load_ntv1(IStream & is, pj_gi_load & gi)
             swap_words(reinterpret_cast<char*>(&row_buf[0]), 8, (int)r_size);
 
         // convert seconds to radians
-        for (boost::int32_t i = 0; i < gi.ct.lim.lam; i++ )
+        for (std::int32_t i = 0; i < gi.ct.lim.lam; i++ )
         {
             pj_ctable::flp_t & cvs = gi.ct.cvs[row * gi.ct.lim.lam + (gi.ct.lim.lam - i - 1)];
 
@@ -327,7 +327,7 @@ inline bool pj_gridinfo_load_ntv2(IStream & is, pj_gi_load & gi)
     std::vector<float> row_buf(r_size);
     gi.ct.cvs.resize(gi.ct.lim.lam * gi.ct.lim.phi);
 
-    for (boost::int32_t row = 0; row < gi.ct.lim.phi; row++ )
+    for (std::int32_t row = 0; row < gi.ct.lim.phi; row++ )
     {
         is.read(reinterpret_cast<char*>(&row_buf[0]), ch_size);
 
@@ -343,7 +343,7 @@ inline bool pj_gridinfo_load_ntv2(IStream & is, pj_gi_load & gi)
         }
 
         // convert seconds to radians
-        for (boost::int32_t i = 0; i < gi.ct.lim.lam; i++ )
+        for (std::int32_t i = 0; i < gi.ct.lim.lam; i++ )
         {
[…]
