### 2026-04-24 bacf0c7
[extensions] Remove dependency on boost/cstdint.hpp
