### 2026-04-24 74ac62e
[extensions] Remove dependency on boost/cstdint.hpp
