+++ include/boost/geometry/extensions/gis/io/shapefile/read.hpp
@@ -11,9 +11,9 @@
 
 
 #include <algorithm>
+#include <cstdint>
 #include <vector>
 
-#include <boost/cstdint.hpp>
 #include <boost/endian/conversion.hpp>
 #include <boost/throw_exception.hpp>
 
@@ -91,7 +91,7 @@ inline void double_endianness_check()
 
     double d = 0;
     unsigned char* c = reinterpret_cast<unsigned char*>(&d);
-    boost::int64_t* i = reinterpret_cast<boost::int64_t*>(&d);
+    std::int64_t* i = reinterpret_cast<std::int64_t*>(&d);
 
     c[0] = 0xd0;
     c[1] = 0x61;
@@ -119,17 +119,17 @@ inline void read_little(IStream & is, double & v)
     static_assert(sizeof(double) * CHAR_BIT == 64, "double must be 64 bits wide.");
 
     is.read(reinterpret_cast<char*>(&v), sizeof(double));
-    boost::int64_t * proxy = reinterpret_cast<boost::int64_t *>(&v);
+    std::int64_t * proxy = reinterpret_cast<std::int64_t *>(&v);
     boost::endian::little_to_native_inplace(*proxy);
 }
 
 template <typename IStream>
-inline boost::int32_t reset_and_read_header(IStream & is)
+inline std::int32_t reset_and_read_header(IStream & is)
 {
     is.clear();
     is.seekg(0);
 
-    boost::int32_t code = 0;
+    std::int32_t code = 0;
     read_big(is, code);
 
     if (code != 9994)
@@ -138,9 +138,9 @@ inline boost::int32_t reset_and_read_header(IStream & is)
     }
 
     // 5 unused, length, version
-    is.seekg(7 * sizeof(boost::int32_t), IStream::cur);
+    is.seekg(7 * sizeof(std::int32_t), IStream::cur);
 
-    boost::int32_t type = 0;
+    std::int32_t type = 0;
     read_little(is, type);
 
     // TODO: support filtering
@@ -173,9 +173,9 @@ inline bool read_record_header(IStream & is)
     //read_big(is, number);
     //read_big(is, length);
 
-    is.seekg(sizeof(boost::int32_t), IStream::cur);
+    is.seekg(sizeof(std::int32_t), IStream::cur);
     // only to set flags
-    boost::int32_t foo;
+    std::int32_t foo;
     read_native(is, foo);
 
     return is.good();
@@ -341,11 +341,11 @@ inline void read_and_set_points(IStream & is, Range & rng, std::size_t num_point
 
 template <typename IStream>
 inline void read_parts(IStream & is,
-                       std::vector<boost::int32_t> & parts,
-                       boost::int32_t num_parts)
+                       std::vector<std::int32_t> & parts,
+                       std::int32_t num_parts)
 {
     parts.resize(num_parts);
-    for (boost::int32_t i = 0 ; i < num_parts ; ++i)
+    for (std::int32_t i = 0 ; i < num_parts ; ++i)
     {
         read_little(is, parts[i]);
     }
@@ -354,10 +354,10 @@ inline void read_parts(IStream & is,
 struct read_point_policy
 {
     template <typename IStream, typename Points, typename Strategy>
-    static inline void apply(IStream & is, Points & points, boost::int32_t type,
+    static inline void apply(IStream & is, Points & points, std::int32_t type,
                              Strategy const&)
     {
-        boost::int32_t t;
+        std::int32_t t;
         read_little(is, t);
         if (t != type)
         {
@@ -386,12 +386,12 @@ struct read_point_policy
 struct read_multipoint_policy
 {
     template <typename IStream, typename Points, typename Strategy>
-    static inline void apply(IStream & is, Points & points, boost::int32_t type,
[…]
