### 2026-04-24 0a3b0e3
Remove dependency on boost/cstdint.hpp
