### 2026-04-24 1f28eb2
Remove dependency on boost/cstdint.hpp
