+++ include/boost/geometry/srs/projections/impl/pj_gridinfo.hpp
@@ -44,7 +44,7 @@
 
 
 #include <algorithm>
-#icnlude <cstdint>
+#include <cstdint>
 #include <string>
 #include <vector>
 
