### 2026-04-25 f03488c
[extensions] Remove dependency on Boost.TypeTraits.
