### 2026-04-25 3d2d373
[extensions] Remove dependency on Boost.TypeTraits.
