+++ include/boost/geometry/extensions/gis/io/shapelib/shape_reader.hpp
@@ -13,8 +13,6 @@
 #include "shapefil.h"
 
 
-#include <boost/type_traits/promote.hpp>
-
 #include <boost/geometry/extensions/gis/io/shapelib/shp_read_object.hpp>
 
 
