### 2026-04-25 81684e8
Remove dependency on Boost.TypeTraits.
