+++ index/test/rtree/rtree_values.cpp
@@ -8,6 +8,7 @@
 // http://www.boost.org/LICENSE_1_0.txt)
 
 #include <memory>
+#include <type_traits>
 
 #include <rtree/test_rtree.hpp>
 
@@ -27,16 +28,16 @@ void check_convertible_to_value(Rtree const& rt, Convertible const& conv)
 {
     static const bool
         is_conv_to_indexable
-            = boost::is_convertible<Convertible, typename Rtree::indexable_type>::value;
+            = std::is_convertible<Convertible, typename Rtree::indexable_type>::value;
     static const bool
         is_conv_to_value
-            = boost::is_convertible<Convertible, typename Rtree::value_type>::value;
+            = std::is_convertible<Convertible, typename Rtree::value_type>::value;
     static const bool
         is_same_as_indexable
-            = boost::is_same<Convertible, typename Rtree::indexable_type>::value;
+            = std::is_same<Convertible, typename Rtree::indexable_type>::value;
     static const bool
         is_same_as_value
-            = boost::is_same<Convertible, typename Rtree::value_type>::value;
+            = std::is_same<Convertible, typename Rtree::value_type>::value;
 
     BOOST_CHECK_EQUAL(is_same_as_indexable, false);
     BOOST_CHECK_EQUAL(is_same_as_value, false);
@@ -52,7 +53,7 @@ void test_pair()
 {
     typedef std::pair<Box, std::size_t> Value;
 
-    typename boost::remove_const<Box>::type box;
+    typename std::remove_const<Box>::type box;
     bg::assign_zero(box);
 
     Value val(box, 0);
@@ -93,7 +94,7 @@ void test_pair_geom_ptr()
 
     typedef std::pair<Box, polygon_t*> Value;
 
-    typename boost::remove_const<Box>::type box;
+    typename std::remove_const<Box>::type box;
     bg::assign_zero(box);
 
     polygon_t poly;
