### 2026-04-25 2a099f2
Remove dependency on Boost.TypeTraits.
