### 2026-04-26 #10328 @ioannisman
In my previous comment, I gave some feedback on the paper, here I share my overall impression and a few comments on the software.

Overall, I think this is a very strong JOSS submission, as Boost.Geometry is a mature, widely adopted, and important geometry library. The repository has some complexity, but this is expected since Boost.Geometry is part of the larger Boost ecosystem, with its conventions, build system, and long development history.

The functionality was verified by running the tests, which provide significant coverage of the codebase, and by running many examples to inspect the functionality in practice. I have used Boost before, but not Boost.Geometry specifically, so I was partly approaching the software as a new Boost.Geometry user.

- Regarding the examples, I agree with @kenohori's point about discoverability. The examples I inspected are useful, but it is not immediately clear how the different example locations relate to each other, or where a new user should start. For example, `example/` seems to contain standalone tutorial-style programs, while `doc/src/examples/` contains documentation-oriented snippets imported into the manual. A short top-level guide explaining these locations and their intended purpose would make the software easier to approach. It would also be useful to polish and unify the examples, and to make their API coverage more systematic.

Regarding my feedback on the paper, I think it would be nice if these points were addressed before acceptance, as this would improve the quality of the paper. Given my feedback on the software, I would definitely not call it a blocker, since this can develop and be improved over time, but I think improving the examples discoverability and consistency could be very useful for new users.
