### 2026-05-05 1a39f4e
Addressing reviewer comments for JOSS
