+++ joss_paper/paper.md
@@ -40,30 +40,33 @@ bibliography: paper.bib
 
 Computational geometry underpins numerous scientific and engineering domains, including GIS, robotics, computer graphics, CAD, and astronomy. Foundational algorithms such as convex hulls, intersection tests, and point-in-polygon queries are covered in standard texts [@deBerg:2008]; however, robust and efficient implementations are crucial for real-world applications where numerical robustness and performance matter.
 
-`Boost.Geometry`[^*] is an open-source C++ library providing generic, extensible, and standards-compliant geometric computation. It defines geometry *concepts* and implements algorithms in a dimension- and coordinate-system-agnostic manner. Since its inclusion in Boost (2011), it has become a widely used component in C++ scientific and industrial software.
+`Boost.Geometry`[^*] is an open-source C++ library providing generic, extensible, and standards-compliant geometric computation. It defines geometry *concepts* and implements algorithms in a dimension- and coordinate-system-agnostic manner. Since its inclusion in Boost (2011), it has been used in a range of C++ scientific and industrial software.
 
 [^*]: Authors are listed in alphabetical order by surname.
 
 # Statement of need
 
-`Boost.Geometry` fulfills the need for a general-purpose, high-performance geometry engine in C++ with:
+C++ developers working in GIS, robotics, computer graphics, CAD, simulation, and scientific computing frequently require geometry algorithms that are both correct and efficient across multiple coordinate systems. Existing solutions either target a single coordinate system, lack extensibility for user-defined types, or impose runtime overhead through dynamic dispatching.
+`Boost.Geometry` addresses these needs by providing:
 
-- A generic programming architecture via concepts, type traits, and template metaprogramming
+- A generic programming architecture via concepts, type traits, and template metaprogramming, enabling zero-overhead abstraction
 - Strategy-based dispatch enabling coordinate-system-specific algorithms
 - Support for Cartesian, spherical, and geographic coordinate systems, including CRS transformations, projections, and spatial indexing
 - OGC-compliant geometry types and spatial predicates [@OGC:2011]
-- Header-only distribution, simplifying deployment and integration
-- Support for adapting user-defined geometry types
+- Header-only distribution, simplifying deployment and integration into existing codebases without additional build dependencies
+- Seamless adaptation of user-defined geometry types without modifying existing data structures
 
-# State of the field 
+# State of the field
 
-A similar library is GEOS [@GEOS:2025], a C++ port of the Java Topology Suite (JTS) that implements the OGC Simple Features standard for GIS applications and serves as the core geometry engine for major geospatial tools like PostGIS [@PostGIS:2025].
+A similar library is GEOS [@GEOS:2025], a C++ port of the Java Topology Suite (JTS) that implements the OGC Simple Features standard for GIS applications and serves as the core geometry engine for major geospatial tools like PostGIS [@PostGIS:2025]. GEOS is widely used in GIS, but it is centered on its own runtime geometry model and API. By contrast, `Boost.Geometry` is designed as a generic C++ library: users can work with adapted user-defined types, write against stateless free functions, and integrate geometry algorithms into existing C++ code without introducing a separate object hierarchy.
 
-CGAL [@Fabri:2000] is a comprehensive C++ library covering a broader range of geometric algorithms (triangulations, Voronoi diagrams, mesh generation, geometry processing) with a strong emphasis on robustness via exact geometric predicates and constructions.
[…]
