### 2026-05-05 f8281fb
Adding explanation about BG examples in README and quick start
