+++ README.md
@@ -17,8 +17,8 @@ Boost.Geometry, part of collection of the [Boost C++ Libraries](http://github.co
 
 ### Directories
 
-* **doc** - QuickBook documentation sources
-* **example** - Boost.Geometry examples
+* **doc** - QuickBook documentation sources, containing embedded examples located in `doc/src/examples/`
+* **example** - Boost.Geometry standalone examples
 * **_extensions_** - examples and tests for the extensions - _develop branch_
 * **include** - the sourcecode of Boost.Geometry
 * **index** - examples and tests for the Spatial Index
