### 2026-05-05 749244d
Qualify sqrt call with math namespace
