### 2026-05-05 d3a4f79
Add curly braces for if statement
