### 2026-05-05 e23b0d7
Improve readability of formulas
