### 2026-05-05 6752c5c
Update test reference value
Area result is now more accurate with more accurate azimuths from andoyer
