### 2026-05-07 905ea42
Add comment for haversine - cosine relation
