### 2026-05-09 adb62e0
Add short distance test
