+++ test/formulas/inverse_short_distance.cpp
@@ -1,0 +1,85 @@
+// Short-distance accuracy test for inverse formulas.
+//
+// Reference for the regression in Andoyer: https://github.com/boostorg/geometry/issues/1217
+// Fix proposed in PR #1461.
+//
+// On develop the pr_1461 case fails: andoyer returns ~9 cm for a true
+// distance of ~0.67 mm. After the PR, andoyer agrees with the reference values
+// from GeographicLib within 1% across the whole range tested here.
+
+#include <geometry_test_common.hpp>
+
+#include <boost/geometry/formulas/andoyer_inverse.hpp>
+#include <boost/geometry/formulas/karney_inverse.hpp>
+//#include <boost/geometry/formulas/thomas_inverse.hpp>
+#include <boost/geometry/formulas/vincenty_inverse.hpp>
+#include <boost/geometry/srs/spheroid.hpp>
+#include <boost/geometry/util/math.hpp>
+
+namespace
+{
+
+struct short_case
+{
+    char const* name;
+    double lon1, lat1, lon2, lat2;
+    double distance_expected;
+};
+
+void test_short(short_case const& c)
+{
+    double const d2r = bg::math::d2r<double>();
+    bg::srs::spheroid<double> const spheroid; // WGS84 by default
+
+    using andoyer_t = bg::formula::andoyer_inverse<double, true, false, false, false, false>;
+    //using thomas_t = bg::formula::thomas_inverse<double, true, false, false, false, false>;
+    using vincenty_t = bg::formula::vincenty_inverse<double, true, false, false, false, false>;
+    using karney_t = bg::formula::karney_inverse<double, true, false, false, false, false>;
+
+    double const lon1r = c.lon1 * d2r;
+    double const lat1r = c.lat1 * d2r;
+    double const lon2r = c.lon2 * d2r;
+    double const lat2r = c.lat2 * d2r;
+
+    double const distance_andoyer = andoyer_t::apply(lon1r, lat1r, lon2r, lat2r, spheroid).distance;
+    //double const distance_thomas = thomas_t::apply(lon1r, lat1r, lon2r, lat2r, spheroid).distance;
+    double const distance_vincenty = vincenty_t::apply(lon1r, lat1r, lon2r, lat2r, spheroid).distance;
+    double const distance_karney = karney_t::apply(lon1r, lat1r, lon2r, lat2r, spheroid).distance;
+
+    double const percent_tolerance = 1.0;  // allow error of 1%
+    BOOST_TEST_INFO_SCOPE(c.name);
+    BOOST_CHECK_CLOSE(distance_andoyer, c.distance_expected, percent_tolerance);
+    //BOOST_CHECK_CLOSE(distance_thomas, c.distance_expected, percent_tolerance); // TODO: Thomas is very inaccurate
+    BOOST_CHECK_CLOSE(distance_vincenty, c.distance_expected, percent_tolerance);
+    BOOST_CHECK_CLOSE(distance_karney, c.distance_expected, percent_tolerance);
+}
+
+} // namespace
+
+int test_main(int, char*[])
+{
+    // reference expected distance values obtained with GeodSolve/GeographicLib
+
+    // Marquee case from PR #1461: ~0.67 mm true distance at mid-latitude.
+    // Develop returns ~9 cm here with Andoyer (the regression the PR fixes).
+    test_short({"pr_1461",         8.81, 53.08, 8.81000001, 53.08, 0.0006701306 });
+
[…]
