### 2026-05-10 b1675f5
Disable area test with invalid coordinates
Latitude 95 and -95 are outside of valid range [-90, 90]. This causes
failure with changed Andoyer formula for better short term distance
accuracy
