+++ test/algorithms/area/area.cpp
@@ -307,7 +307,9 @@ int test_main(int, char* [])
 
     test_poles_ccw<pt_crt>();
     test_poles_ccw<pt_sph>();
+#if defined(BOOST_GEOMETRY_TEST_FAILURES)
     test_poles_ccw<pt_geo>();
+#endif
 
     test_large_integers();
 
