### 2026-05-10 43cbd7c
Add explanation comment for disabled test
