+++ test/algorithms/area/area.cpp
@@ -307,7 +307,7 @@ int test_main(int, char* [])
 
     test_poles_ccw<pt_crt>();
     test_poles_ccw<pt_sph>();
-#if defined(BOOST_GEOMETRY_TEST_FAILURES)
+#if defined(BOOST_GEOMETRY_TEST_FAILURES) // fails due to invalid coordinates, see PR #1461
     test_poles_ccw<pt_geo>();
 #endif
 
