### 2026-05-19 #10328 @hugoledoux
@editoralbot check references
