### 2026-05-19 5a58f6b
Fix two small typos + protect words in bibtex
