### 2026-05-21 f93f6c7
Add citation info and link to JOSS paper
