### 2026-06-06 b8246cc
fix: failing tests on macos in release mode
Implements: #1464
