### 2026-06-06 a9660c9
fix: failing tests on macos in release mode
Implements: #1464
