### 2026-06-14 adf0d00
Fix degenerate linestring relation on boundary of areal
