+++ include/boost/geometry/algorithms/detail/relate/linear_areal.hpp
@@ -86,6 +86,11 @@ public:
         {
             m_interrupt_flags |= 8;
         }
+
+        if ( ! may_update<interior, boundary, '1', TransposeResult>(m_result) )
+        {
+            m_interrupt_flags |= 0x10;
+        }
     }
 
     template <typename Linestring>
@@ -102,7 +107,7 @@ public:
         }
 
         // if those flags are set nothing will change
-        if ( m_interrupt_flags == 0xF )
+        if ( m_interrupt_flags == 0x1F )
         {
             return false;
         }
@@ -110,12 +115,19 @@ public:
         int const pig = detail::within::point_in_geometry(range::front(linestring),
                                                           m_geometry2,
                                                           m_strategy);
-        //BOOST_GEOMETRY_ASSERT_MSG(pig != 0, "There should be no IPs");
 
         if ( pig > 0 )
         {
             update<interior, interior, '1', TransposeResult>(m_result);
             m_interrupt_flags |= 1;
+        }
+        else if ( pig == 0 )
+        {
+            // no turns but still point on boundary can actually happen when the
+            // linestring is degenerate. so handle this case explicitly here.
+            // for consistency let's still report dimension 1 instead of 0
+            update<interior, boundary, '1', TransposeResult>(m_result);
+            m_interrupt_flags |= 0x10;
         }
         else
         {
@@ -140,7 +152,7 @@ public:
             }
         }
 
-        return m_interrupt_flags != 0xF
+        return m_interrupt_flags != 0x1F
             && ! m_result.interrupt;
     }
 
