### 2026-06-17 80cd1f6
Use hexadecimal for all interrupt flags
