### 2026-06-17 4ff4b94
Also support single-point linestrings
