### 2026-06-28 b977460
fix: repair dissolve for new traversal
