+++ include/boost/geometry/algorithms/detail/overlay/graph/traverse_graph.hpp
@@ -190,17 +190,17 @@ struct traverse_graph
         // more than times there are turns (this should not happen).
         while (iteration_count < m_turns.size())
         {
-            auto const current_turn_indices = get_turn_indices_by_node_id(m_turns, m_clusters,
-                    current_node_id, allow_closed);
+        auto const current_turn_indices = get_turn_indices_by_node_id(m_turns, m_clusters,
+                current_node_id, allow_closed);
 
-            // Any valid node should always deliver at least one turn
-            BOOST_ASSERT(! current_turn_indices.empty());
+        // Any valid node should always deliver at least one turn
+        BOOST_ASSERT(! current_turn_indices.empty());
 
-            auto const next_target_nodes = get_target_nodes<target_operation>(m_turns, m_clusters,
-                    current_turn_indices, component_id);
+        auto const next_target_nodes = get_target_nodes<target_operation>(m_turns, m_clusters,
+                current_turn_indices, component_id);
 
-            if (next_target_nodes.empty())
-            {
+        if (next_target_nodes.empty())
+        {
 #if defined(BOOST_GEOMETRY_DEBUG_TRAVERSE_GRAPH)
                 std::cout << "Stuck, start: " << start_node_id
                     << " stuck: " << current_node_id
@@ -209,8 +209,8 @@ struct traverse_graph
                 return false;
             }
 
-            auto const tois = get_tois<target_operation>(m_turns, m_clusters,
-                    current_node_id, next_target_nodes);
+        auto const tois = get_tois<target_operation>(m_turns, m_clusters,
+                current_node_id, next_target_nodes);
 
             if (tois.empty())
             {
@@ -221,12 +221,12 @@ struct traverse_graph
 
             auto toi = *tois.begin();
 
-            if (tois.size() > 1)
-            {
-                // Select the best target edge, using the last point of the ring and the turn point
-                // for side calculations (if any).
+        if (tois.size() > 1)
+        {
+            // Select the best target edge, using the last point of the ring and the turn point
+            // for side calculations (if any).
                 toi = m_edge_selector.select_target_edge(tois, range::back(ring), turn_point);
-            }
+        }
 
             if (m_visited_tois.count(toi) > 0 || m_finished_tois.count(toi) > 0)
             {
@@ -272,7 +272,7 @@ struct traverse_graph
         // Select the first toi which is not yet visited and has the requested component.
         // If all tois are visited, not having the same component, it is not possible to continue,
         // and it returns an invalid toi.
-        auto select_first_toi = [&](auto const& tois)
+        auto select_first_toi_other = [&](auto const& tois)
         {
             for (auto const& toi : tois)
             {
@@ -290,6 +290,40 @@ struct traverse_graph
                 return toi;
             }
             return turn_operation_id{0, -1};
+        };
+
+        auto select_first_toi_of_two_in_union = [&](auto const& tois)
+        {
+            const auto& toi0 = *tois.begin();
+            const auto& toi1 = *(++tois.begin());
+
+            if (m_finished_tois.count(toi0) == 0
+                && m_finished_tois.count(toi1) == 0)
+            {
+                auto const& turn0 = m_turns[toi0.turn_index];
+                auto const& turn1 = m_turns[toi1.turn_index];
+                auto const& op0 = turn0.operations[toi0.operation_index];
+                auto const& op1 = turn1.operations[toi1.operation_index];
+
+                if (op0.preference_index != op1.preference_index) 
+                {
+                    return op0.preference_index < op1.preference_index
+                        ? toi0
+                        : toi1;
+                }
+            }
+            return select_first_toi_other(tois);
+        };
+
+        auto select_first_toi = [&](auto const& tois)
+        {
+            if (tois.size() == 2
+                && target_operation == operation_union)
[…]
