### 2026-06-29 d36b9a1
fix: degenerate linestring relation on boundary of areal
Fixes: #1474

* Fix degenerate linestring relation on boundary of areal
* Use hexadecimal for all interrupt flags
* Also support single-point linestrings
