### 2026-06-29 ed87d43
fix: nearly identical input should give an intersection
Fixes: #1471
