+++ include/boost/geometry/algorithms/detail/overlay/select_rings.hpp
@@ -238,7 +238,8 @@ inline void update_ring_selection(Geometry1 const& geometry1,
             TurnInfoMap const& turn_info_map,
             RingPropertyMap const& all_ring_properties,
             RingPropertyMap& selected_ring_properties,
-            Strategy const& strategy)
+            Strategy const& strategy,
+            bool include_on_boundary = false)
 {
     selected_ring_properties.clear();
 
@@ -261,22 +262,29 @@ inline void update_ring_selection(Geometry1 const& geometry1,
             continue;
         }
 
-        // Check if the ring is within the other geometry, by taking
-        // a point lying on the ring
+        // Check if the ring is within the other geometry, by using
+        // a point on the ring.
+        int code = 0;
         switch(id.source_index)
         {
             // within
             case 0 :
-                info.within_other = range_in_geometry(pair.second.point,
-                                                      geometry1, geometry2,
-                                                      strategy) > 0;
+                code = range_in_geometry(pair.second.point,
+                                         geometry1, geometry2, strategy);
                 break;
             case 1 :
-                info.within_other = range_in_geometry(pair.second.point,
-                                                      geometry2, geometry1,
-                                                      strategy) > 0;
+                code = range_in_geometry(pair.second.point,
+                                         geometry2, geometry1, strategy);
                 break;
         }
+
+        // Code 0: the point is located on the boundary. Normally such a
+        // situation is resolved by the turns (the rings are traversed).
+        // But if the overlay produced no turns at all, and the inputs are (nearly)
+        // coincident without any segment intersection being detected (such as
+        // issue 1471), then a point on the boundary should be counted as being
+        // within the other geometry.
+        info.within_other = include_on_boundary ? code >= 0 : code > 0;
 
         if (decide<OverlayType>::include(id, info))
         {
@@ -303,7 +311,8 @@ template
 inline void select_rings(Geometry1 const& geometry1, Geometry2 const& geometry2,
             RingTurnInfoMap const& turn_info_per_ring,
             RingPropertyMap& selected_ring_properties,
-            Strategy const& strategy)
+            Strategy const& strategy,
+            bool include_on_boundary = false)
 {
     using tag1 = geometry::tag_t<Geometry1>;
     using tag2 = geometry::tag_t<Geometry2>;
@@ -318,7 +327,7 @@ inline void select_rings(Geometry1 const& geometry1, Geometry2 const& geometry2,
 
     update_ring_selection<OverlayType>(geometry1, geometry2, turn_info_per_ring,
                 all_ring_properties, selected_ring_properties,
-                strategy);
+                strategy, include_on_boundary);
 }
 
 template
@@ -332,7 +341,8 @@ template
 inline void select_rings(Geometry const& geometry,
             RingTurnInfoMap const& turn_info_per_ring,
             RingPropertyMap& selected_ring_properties,
-            Strategy const& strategy)
+            Strategy const& strategy,
+            bool include_on_boundary = false)
 {
     RingPropertyMap all_ring_properties;
     dispatch::select_rings<geometry::tag_t<Geometry>, Geometry>::apply(geometry,
@@ -341,7 +351,7 @@ inline void select_rings(Geometry const& geometry,
 
     update_ring_selection<OverlayType>(geometry, geometry, turn_info_per_ring,
                 all_ring_properties, selected_ring_properties,
-                strategy);
+                strategy, include_on_boundary);
 }
 
 
