### 2026-06-29 a7875f6
fix: nearly identical input should give an intersection
Fixes: #1471
